init: Mount stateful filesystem and dev-image with discard
In preparation for use with thin provisioned logical volumes,
mount the stateful filesystem and the dev-image filesystem with
discard. This allows the space cleaned up by these filesystems
to be reclaimed by the underlying thinpool. An additional benefit
is that this renders chromeos-trim redundant.

BUG=b:200582822
TEST=install and check mount options for stateful and dev-image

Change-Id: If14efe0f956408771795e362ffcd4ccb8bb8d1ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3171205
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>