chromeos-config: Initial copy of proto converter
Straight copy from src/config/payload_utils with the unittest disabled
for now.

In order to move to dynamic config compilation/transforms in the ebuild,
this logic needs to be part of cros_config_host, so moving there so it
can be installed as part of the host sdk package.

Next CL will be getting this working in cros_config_host

BUG=b:195298103
TEST=None

Change-Id: Ic1c6d4ccf6b598fc3a7930ac174ba388e15b39a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3174074
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Commit-Queue: C Shapiro <shapiroc@chromium.org>
Tested-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>