cryptohome: [USS] Basic storage implementation
Implement a basic file storage for the UserSecretStash: writing it to a
file and loading it back. A basic robustness guarantee is achieved using
fdatasync+fsync (the former for the file contents, the latter - for the
parent directory).

The future plan will be to increase the robustness further by
implementing versioning, i.e., keeping a few most recent files with the
possibility to roll back to the last non-corrupted one.

BUG=b:200147797
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: I6e2bdbd63bc937b7d728a8bb8691a465d663e55e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3173361
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Auto-Submit: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>