cryptohome: [USS] Support loading via wrapping key
Add a new way of constructing the UserSecretStash, given the USS
container flatbuffer and the wrapping key. For example, this is to be
used when the user authenticates using a credential and cryptohome wants
to load&decrypt the USS using the output of the auth block.

This method of construction completes two other methods of construction,
which are (a) random initialization - to be used when creating a new
USS, and (b) loading via the USS Main Key - to be used when there's a
previously authenticated AuthSession.

BUG=b:201144150
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: Ib9071c28e8f35128db4c189b80ada4d2b5373f7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3176240
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>