cryptohome: Unit tests for AddKeyset overloaded to work with AuthSessions
Adds 5 unit tests for the overloaded KeysetManagement::AddKeyset.
One test an open path is found, and enters a new label. Another
similar test, but the label already exists. There are two tests
where the open path exists, but either encryption or save fails.
Another tests for VaultKeyset keeping the existing keyset valid
on disk when it tries overwriting an existing label and fails.

Also changes an OpenFile() call in AddKeyset to use the
platform.h wrapper, for consistency with other calls in the
class.

BUG=b:202315245
TEST=FEATURES=test emerge-${BOARD} cryptohome

Change-Id: I6e7b5c113946e0e3a14a7df6ccbc0e3a97fed30a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3177201
Tested-by: Thomas Cedeno <thomascedeno@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Thomas Cedeno <thomascedeno@google.com>