ml: Annotate intentionally leaked model pointers for HWR, WebHWR, SODA.
In production, model pointers (for process-isolated models) are
intentionally leaked. This is because the model runs in its own process
and the model's memory is freed when the process exits.

This intentional leak can cause problems with ASAN testing, so we
annotate these intentional leaks as such.

Note that this is being done pre-emptively / as good practice in this
CL. The absence of these annotations was not causing problems for
existing tests for these three models.

BUG=chromium:1220271
TEST=CQ

Change-Id: Ib81f890a35521940c22f4bedd2a79b287581c9fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3176830
Reviewed-by: Honglin Yu <honglinyu@chromium.org>
Commit-Queue: Amanda Deacon <amandadeacon@chromium.org>
Tested-by: Amanda Deacon <amandadeacon@chromium.org>