secanomalyd: Extract mount-related operations to a file.
We're going to need to generate a list of mounts from several places in
the code, so extract this functionality to a separate file. This also
allows for easier unit testing, so add some unit tests for our mount
reading and parsing code.

BUG=b:200991907
TEST=New unit tests.
TEST=Deploy, run secanomalyd, mounts are still identified.

Change-Id: Icedb95ea7fdc6f41af95342f91277572ccf2f1e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3182704
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Nicole Anderson-Au <nvaa@google.com>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>