shill: vpn: Report PPPAuth failure in L2TPConnection
This patch adds the logic to report kFailurePPPAuth instead of
kFailureInternal when the connection failure is caused by authentication
error. It is implemented by detecting the exit code of pppd. To do this,
this patch adds a kPPPReasonExit message in the Notify() interface to
forward the exit code.

Note that in some other usages of pppd in shill, the authentication
failure has already been treated differently based on the exit code
(e.g., in cellular.cc). In those places, it is implemented by
registering the death callback when starting pppd. We cannot do the same
thing in L2TPConnection, since the pppd process is not started by shill
directly, but by xl2tpd instead, and thus we need the IPC channel to
forward the information back to shill.

Other notable changes:
- In PPPDevice::ExitStatusToFailure(), also map EXIT_AUTH_TOPEER_FAILED
  (which is 19) to kFailurePPPAuth. This exit code will be use when we
  have the wrong username/password on DUT.
- Make sure other pppd users in shill ignore this kPPPReasonExit
  message.

BUG=b:165170125
TEST=manually checked that UI showed "PPP authentication error" when
  using a wrong username or password.

Change-Id: Ib0325308f85a0e5634c5df9492c389fb5cd4c141
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3185101
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>