libbrillo: migrate DevmapperTaskFactory to base::RepeatingCallback.
dm_task_factory_ in device_mapper.h is a DevmapperTaskFactory. It is
used in DeviceMapper::Setup(), DeviceMapper::Remove(), and so on, and
from the unit tests, for example FakeTaskConformance, we see that
dm_task_factory_ will be called multiple times, so we need a
RepeatingCallback. The same conclusion can be reached by analyzing
DmcryptContainer, where calls to Setup and Remove have the same effect.

The main goal of this change is to migrate usage of base::Bind in
cryptohome to base::BindRepeating.

BUG=chromium:1205291
TEST=USE=test emerge-betty libbrillo cryptohome

Change-Id: I5ebfd392969fbc0e5b4a39fe6fd62660e531ad21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3185109
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>