hps: Add a framework for filtering results from HPS.
Introduce the concept of a Filter for HPS inferencing results. Filters
take the results of a single HPS inference and determine if this
inference result is valid (true) or invalid (false).

To support signaling when the filter state changes, we introduce a
FilterWatcher that is composed of another filter. When that Filter
changes the inference result from what has previously been observed,
then FilterWatcher will invoke a callback so that a signal can be
raised.

We also introduce basic tests for filters.

Further work here includes:
- Introducing mode complicated filters.
- Better error handling for filters when HPS is returning errors.
- Teasing apart the HPS result from an int into it's parts so it's
easier to consume.

BUG=b:198682678
TEST=FEATURES=test emerge-brya chromeos-base/hpsd

Cq-Depend: chromium:3184075
Change-Id: I62cc43f7b1167ba447cdb21560347cbae50d517e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3185107
Tested-by: Stuart Langley <slangley@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Reviewed-by: Evan Benn <evanbenn@chromium.org>