shill: vpn: Fix issue in the new L2TPIPsecDriver on irregular events
Current, the new L2TPIPsecDriver does not invoke the disconnect callback
on suspend or default service changed events, and thus the service will
show "Connected" in the UI when those events happen (it is actually
disconnected). The reason is that currently the driver calls
VPNConnection::Disconnect() on those events, but VPNConnection will not
invoke any callback when its state becomes disconnecting. So this patch
adds the logic to trigger NotifyServiceOfFailure() explicitly when those
events happen.

BUG=b:165170125
TEST=Checked that the service showed "disconnected" on suspend / default
  physical service changed or down.

Change-Id: I3009149a4c0a79fe4f84543b84c602eda33d98a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3185499
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>