cryptohome: Use RepeatingCallbacks in StatefulRecovery.
StatefulRecovery defines 3 callback types: MountFunction,
UnmountFunction, and IsOwnerFunction.

RepeatingCallbacks are more suitable for this use case, because the
functions are passed into StatefulRecovery as a way to abstract away the
details of calling UserDataAuth, and not as a callback to events, and
therefore, StatefulRecovery is actually allowed to call these more than
once.

BUG=chromium:1205291
TEST=USE=test emerge-betty cryptohome

Change-Id: I0d59483dccfbe486881262f4ca67ce93a52e0213
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3190065
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>
Auto-Submit: Tomasz Tylenda <ttylenda@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>