cryptohome: Simulate mounts and binds in the fake platform
The mounts and binds are very simplistic now and don't understand any
semantics of the various mount modes, but that will be adjusted as we
rework mount_unittest.cc

The internal semantics of mode/owner/xattr also changes to account for
the possibility of representing the same file under multiple paths in
the case of bind. The mode and ownership routines now also check for the
file presence, which was blocked by the lack of fake mounts (the exact
reason why so is a bit complex and let me not consume too much space
with an already large commit message).

The fake mapper comes with extensive unittest, and
encrypted_fs_unittest.cc is the first consumer of the functionality
added with the FakeMountMapper.

BUG=b:172622448
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: I27875c99ed29370d247d9890236cc6cdb99581cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3194214
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>