camera: teach Gcam AE to handle the request and result metadata
For Gcam AE to act as the main camera AE controller, we need to handle
and report the request and result metadata following the API contract.

BUG=b:201596601
TEST=Manually on DUT and verify that 1) AE states and face ROIs are
shown in CCA's expert mode, and 2) the we get 15-30 fps in picutre mode
and constant 30 fps in video mode.

Change-Id: I0bd3f8573e0560618ac9175cde5799f8943d8516
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3193340
Reviewed-by: Wei Lee <wtlee@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>