lorgnette: Merge usb and net test devices
Some MFPs have a hex number in brackets at the end of the name they
advertise on the network.  The USB version of the name will be identical
minus the trailing digits.  Since they are the same device, having to
keep track of which name can be used with each connection type has
turned out to be confusing in our test plan.

Remove the trailing digits so that the network name merges with the USB
name.  This lets testers treat these devices the same as others that
don't have this behavior.  Since the mDNS name is only used for display,
this has no effect on test behavior.

BUG=b:201081168
TEST=set_wwcb_mfp_connection_type finds USB and net for the same printer

Change-Id: I5dc8ffff342567af8ba46e5c935b1e23c2e31c86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3194570
Reviewed-by: Paul Moy <pmoy@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>