shill: cellular: Use loose comparison for APNs
When comparing APNs to be used for connection, ignore APN properties
that are not related to the connection. This will allow us to ignore
APNs that are the same, but have properties that were added during
runtime, and will be used to better know the origin of the APN when
logging metrics.

BUG=b:197020650
TEST=Run multiple connection scenarios using the UI, and verify the
the APN used when connecting.

Change-Id: I5e7e6d31f3e934fe90b5689668c722392fb19484
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3198311
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>