cryptohome: remove an unnecessary pkcs11 state
That if can never be true. The only place where the state is set to
kIsBeingInitialized is in [1], and it will unconditionally transition to
kIsInitialized at [2]. Since both [1] and [2] and the function with the
"if" are asserted to be executed on the mount thread, there is no way for
it to interject between [1] and [2], and thus the condition can never
happen. More over, with the condition gone, there is no other code
referring to the state, which mean the state is unnecessary.

This is a part of a cleanup to move parts of Pkcs11 handling from Mount

[1] https://source.chromium.org/chromium/chromiumos/platform2/+/HEAD:cryptohome/userdataauth.cc;l=962;drc=79cd226092bffb047487a29c3b8fa2a4a5704204
[2] https://source.chromium.org/chromium/chromiumos/platform2/+/HEAD:cryptohome/userdataauth.cc;l=971;drc=79cd226092bffb047487a29c3b8fa2a4a5704204

BUG=b:172344610
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: I69f9f75d4f2b7ff5391f3615751abb1aa756b7a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3198915
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>