cryptohome: remove checks for non-mount thread in pkcs11 init flow
InitializePkcs11 and ResumeAllPkcs11Initialization are guaranteed to be
called on the mount thread because all call sites are preasserted to be
on it ([1] for ResumeAllPkcs11Initialization and [2]-[4] for
InitializePkcs11). And to enforce nothing else can unwarrantly call into
InitializePkcs11, move to private, since it is not a part of the public
interface anyway. That would require to make couple of tests friends,
but that is better that keep it around as public just for the sake of
it.

[1] https://source.chromium.org/chromium/chromiumos/platform2/+/HEAD:cryptohome/userdataauth.cc;l=1070;drc=79cd226092bffb047487a29c3b8fa2a4a5704204
[2] https://source.chromium.org/chromium/chromiumos/platform2/+/HEAD:cryptohome/userdataauth.cc;l=989;drc=79cd226092bffb047487a29c3b8fa2a4a5704204
[3] https://source.chromium.org/chromium/chromiumos/platform2/+/HEAD:cryptohome/userdataauth.cc;l=1006;drc=79cd226092bffb047487a29c3b8fa2a4a5704204
[4] https://source.chromium.org/chromium/chromiumos/platform2/+/HEAD:cryptohome/userdataauth.cc;l=1866;drc=79cd226092bffb047487a29c3b8fa2a4a5704204

BUG=b:172344610
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: Ia4dccd04192c92a44d1e7fbccbc997b428a42017
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3198919
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>