cryptohome: remove kIsWaitingOnTPM state
The state is not specifically different from Uninitialized, except that
it already did come to the point of trying to initialize and observing
absence of tpm ownership. However, there are only two paths to get to
initialization: one straight after the mount, without leaving the
thread [1], from the ownership callback [2] and from the restore
callback [3]. This means, that once mounted, the state will immediately
transition to either kIsWaitingOnTPM or kInitialized. In that case, it
makes no real difference to distinguis kIsWaitingOnTPM from
kUninitialized.

[1] https://source.chromium.org/chromium/chromiumos/platform2/+/HEAD:cryptohome/userdataauth.cc;l=1866;drc=79cd226092bffb047487a29c3b8fa2a4a5704204
[2] https://source.chromium.org/chromium/chromiumos/platform2/+/HEAD:cryptohome/userdataauth.cc;l=989;drc=79cd226092bffb047487a29c3b8fa2a4a5704204
[3] https://source.chromium.org/chromium/chromiumos/platform2/+/HEAD:cryptohome/userdataauth.cc;l=1006;drc=79cd226092bffb047487a29c3b8fa2a4a5704204

BUG=b:172344610
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: I08a7603fc02515e543ab805281be24678bee5e3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3198921
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>