secanomalyd: Add process information gathering code.
For the purpose of reporting anomalous system state we're going to need
to report the list of processes on the system.

Also fix up a couple of nits in the code.

This code is currently only exercised by unit tests, but will be used
in the next CL.

BUG=b:200991907
TEST=New unit tests.

Change-Id: I5f3d0fc8c62b6fa66399448975c36cd7d2c36cc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3198747
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>