init: Corral stateful writes on hibernate resumes
A hibernated system maintains active mounts on RW file systems like the
stateful partition. To keep a hibernate image viable to resume, we
cannot modify on-disk structures in between the time the hibernate image
is snapped, and when it is resumed and running.

The "resume boot" is a fresh boot responsible for booting to login,
collecting the user's password, loading the hibernate resume image, and
launching it. The trick is to do all of this without modifying the
stateful partition.

Enter dm-snapshot. We can set up a snapshot image pointed at the
stateful partition, and mount that instead. Writes to the dm-snapshot
can be redirected to a loop device file in tmpfs (RAM). On a successful
hibernate resume, these writes are discarded as if the resume boot never
happened. On a failed hibernate resume, these changes are committed back
down to the stateful partition and become the "new timeline".

This change test if the hibernate service is present, and if so, checks
if the hibernate cookie is set indicating a valid resume image. If it's
set, it sets up the dm-snapshot partition and diverts stateful to there.

BUG=b:202060181
TEST=Test hibernate flow on volteer, observe no FS inconsistencies.

Change-Id: I0cba489a02372c659b5de1b6972d482433c1e449
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3202987
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>