init: Delineate termination for periodic scheduler and children
- periodic_scheduler does not actually kill its child processes
on timeout. Add termination for child process: p.Terminate()
sends SIGTERM to the child process, waits and then sends a
SIGKILL.

- Change PDEATHSIG to SIGKILL so that a crash in the periodic
scheduler immediately kills the child processes and there isn't
a possibility of these processes getting reparented to init.

- Add a SIGTERM handler for propagating the signal to the running
child process. Once the signal is sent, wait for the process to
exit gracefully with a timeout. At the end of the timeout,
the periodic scheduler exits and the child process is sent a
PDEATHSIG (SIGKILL).

BUG=b:197009375
TEST=compilation

Change-Id: I1a0ab44e3588d6abf1773080a5b4462f3e687ce0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3203321
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>