mount-passthrough: Increase open fd limit for files shared via virtio-fs
Currently, the maximum number of open file descriptors of
mount-passthrough processes is set to 8192. On the other hand, virtio-fs
devices can open a lot of fds at the same time (up to 262144, which is
the number defined in vm_concierge.conf).
As a result, we often encounter the "too many open files" error e.g.
when we let a guest process open many (> 8192) files that are shared
via mount-passthrough + virtio-fs.
Here, rather than imposing a new restriction on the number of fds that
virtio-fs devices can open, it would be desirable to just let virtio-fs
devices open as many files as they like, and allow them to manage the
opened files on their own.
In order to achieve that, for mount-passthrough processes that are used
in combination with virtio-fs, increase the maximum number of open fds
to 262144 (i.e. the maximum number of fds virtio-fs devices can open).

BUG=b:201021784
TEST=for i in $(pgrep mount-pass); do
TEST=  echo $(tr '\000' ' ' < /proc/${i}/cmdline)
TEST=  grep 'open files' /proc/${i}/limits
TEST=done

Cq-Depend: chromium:3207708
Change-Id: I65e519e305e0ee81ba38634aea3061ff59306bae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3205227
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>