cryptohome: Make OOP and non-OOP epehemeral teardown similar
So we can remove effectively meaningless at this point teardown in Mount
itself. Note, that we also remove reporting on error, but it is no
behaviour change for the code path is already not excercised. We need to
introduce the check back on the present paths later.

BUG=b:172344610
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: I5eb14f4a5f39ba938da7cc33abd0aad11a8abb99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3211309
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>