shill: cellular: Ensure that the modem is set to low power mode on stop
When we want to set the modem to low power after disabling the modem,
ensure that we do it even if disabling the modem fails. This is because
the modem might be in failed state, but might have the radio ON.
Also, fix unit test that was broken because ReleaseProxies was not being
called when Stop_DisableCompleted was returned with an error.

BUG=b:192634026
TEST=Verify that the modem goes to low power when suspending the device
without a SIM card in it.

Change-Id: I38ab8f501cb04674d6736d2b38934a8d83f586a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3213438
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>