hermes: Skip notifications after profile operation
Sometimes, MM fails to open a channel when Hermes is active. Until the
issue is root caused and fixed, do not acquire a channel after a profile
operation. MM is known to be active during this time and needs to have
higher priority for opening the channel. Thus, skip sending
notifications since they are optional by the standard until we can fix
the issue.

BUG=b:202401139, b:195589882
TEST=tast run ${cr} network.HermesMultiProfile

Change-Id: Icae19d9ccee388963ee8a80b60aff27beaf908cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3212749
Tested-by: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Pavan Holla <pholla@google.com>