cryptohome: Remove two duplicate unit tests from mount
MountTest::MountCryptohomeHasPrivileges is an artifact from the time
when mount.cc was dealing with the keys and had some privilege check.

EphemeralSystemTest::RegularMount (was named differently before) is an
articfact from the time when mount.cc was deciding on its own whether it
needs to mount a user regularly or as an ephemeral.

With those two parts being gone (completely for the first one and out of
mount to the higher level for the second), now those two tests are
correspondingly covered by MountTest::MountCryptohome and
MountTest::MountPristineCryptohome, thus they are unnecessary.

This CL also modifies those two test to do unmount as well to preserve
the full coverage the duplicated tests provided.

BUG=b:172344610,b:172622448
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: I04f5a55b4fb0e6fc1879e834594114429155cbff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3216066
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>