cryptohome: untie ephemeral tests from non-ephemeral
Make the ephemeral tests a thing on its own, rather a child of
MountTest. That allows them to avoid duplicated running for different
backing encryption type (which was meaningless because it does not use
any encryption on ramfs) and will let us further independently refactor
this part for a use with FakePlatform.

The untying requires some pieces of code to be duplicated (setup and
daemon store expectation), but that is going to be reworked in the
following CLs which will aim to remove most mocks from epehemeral test.
The regular mount tests will be refactored once ephemeral path is fixed.

BUG=b:172344610,b:172622448
TEST=FEATURES="test" USE="test" emerge-hatch cryptohome

Change-Id: I17a5495655acedef724fbbba73bff1b17d5fe6cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3216072
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>