chromeos-config: Split out the identity table from configfs
Move some code around so that the identity table generator is
decoupled from the squashfs.  The new output path is from a command
line flag instead of inside of the squashfs.

Note that this deletes the original identity table unit tests that go
in-depth into the table contents.  The crosid e2e tests do it much
better, so we just do a basic check that the size looks good and let
the crosid tests go into the file format.

BUG=b:202766035
TEST=unit tests

Change-Id: I524ec5af65c6de1a41293f1e875fad5fc5bdedcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3217956
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>