secanomalyd: Add unit tests for brillo::Process-related code.
Add unit tests to the code using brillo::Process. In order to do this
we pass a Process pointer to the reading function. This function is not
expected to be used outside of testing, so using a raw pointer should
be fine. The functions used by the rest of the code wrap this in a
unique_ptr.

Also add a few unit tests for other failure cases.

BUG=b:200991907
TEST=New unit tests.

Change-Id: Iff45a88f87ec90e5d7bcf416b5719cce875f8b0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3218331
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>