ml: fix test case with restricted experiment group
The most recent update to libsuggest (see the pending CL
crrev/c/3218925) updates the TextSuggester experiment group
GboardRelaxedA to use stricter thresholds. As a consequence the
suggestions returned from the group are different.

The current integration test for the GboardRelaxedA experiment group
uses a suggestion that will no longer be emitted with the new libsuggest
version.

This CL then anticipates this change by updating the integration test to
use a suggestion that is expected by both the old and new versions of
libsuggest.

BUG=chromium:1146266
TEST=cros_run_unit_tests --board=${BOARD} --packages chromeos-base/ml

Change-Id: If8eec752cef8e94fea4ba185df3f50ad6c783786
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3220192
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Tested-by: Curtis McMullan <curtismcmullan@chromium.org>
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>