shill: persist Passpoint credentials in Profile.
Passpoint credentials are now persisted in the Profile when added
through the Manager API. A set of credentials is identified by an unique
identifier generated at creation. On profile load, the credentials have
to be extracted before the (WiFi) services as a service may reference a
set of credentials.

BUG=b:162106001
TEST=shill unit tests; insert some credentials with "gdbus call" and
  check with traces the credentials are saved correctly, loaded on
  profile push and unloaded on profile pop.

Change-Id: I1b7799f4a843d49e7e9c76ab554e6eed3b2fd8ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3241424
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>