camera: avoid setting request settings when it's null
Cache the latest set of request settings in the camera device adapter
and always pass the request settings to the vendor camera HAL and stream
manipulators.

Only change the metadata in the stream manipulators when the client
request settings is not null. This is mainly to make the CTS tests
happy, as some test cases set null settings and if we change that the
vendor camera HAL may not handle the incremental changes well.

BUG=b:203028235
TEST=run cts-camera -m CtsCameraTestCases

Change-Id: I094f3a4114f5757d248a39d85ed229e9a8d85753
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3227510
Reviewed-by: Wei Lee <wtlee@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>