vm_tools: maitred: switch to base::OnceCallback
Replace the use of the deprecated Callback with OnceCallback for the
Spawn function's exit callback, which is called once when the process
exits. The ResizeFilesystem function is the only user of this optional
parameter, so update that call site to use BindOnce as well.

Additionally use OnceCallback for the shutdown callback, which is called
once when the shutdown RPC is received.

BUG=chromium:1205291
TEST=Deploy maitred to termina and resize disk

Change-Id: I94e7a99bcbc0fbe32b5b11254d6337af54316442
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3221728
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>