lorgnette: Implement more page size options
Several backends offer page-width and page-height options in addition to
the usual scan area.  For these backends, the scan area values are
clamped to the current values of page-width and page-height.  The real
maximum scannable area is given by the ranges of page-width and
page-height.

If a backend exposes these options, use them to calculate the scannable
area instead of the previous values.  This allows the UI to show sizes
larger than US Letter (the default page size).  In addition, set
page-width and page-height to to user's requested page size before
attempting to set the scan area.  This ensures that we don't request a
scan area that's out of range for the current page size.

BUG=b:196357248
TEST=Scan A4 and Legal on an ix500; tast run scanner.*

Change-Id: I4607207ce0749d6e17889f477744b8d12763bbd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3224260
Reviewed-by: Paul Moy <pmoy@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>