shill: Add UMA metric to track network instability after rekey
In b:186763776, we notice that the device's network connection becomes
unstable quickly after a rekey. We do not know the scope of how
prevalent this issue is, so we add a metric to track how often it is
that a device's connection becomes unstable after a rekey attempt.

Once a rekey is initiated, if the network becomes unstable within 3
minutes, we record the number of seconds it took for the device to
fail.

BUG=b:186763776,b:172225523,b:193155280
TEST=shill unit tests

Change-Id: Id1617ccb7c98d3072fa6fd5fc3bf2beb7481a1fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3224374
Tested-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jun Yu <junyuu@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>