debugd: Simplify storage device detection
Use rootdev() to detect the currently booted device that needs
to be diagnosed. This allows us to support logical-volume-based
stateful partitions. Additionally, this change removes the need
to parse the mount table and so we can remove unittests that just
test the mount parsing.

BUG=b:172344853
TEST=health.DiagnosticsRun.nvme_self_test passes on an LVM stateful
partition.

Disallow-Recycled-Builds: test-failures
Change-Id: I8f70d1381f4f22833012edc10dc4111195316798
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3224564
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>